ml_Solutions/vertex_ai/app-engine-deploy/requirements.txt
streamlit
google-cloud-aiplatform
pandas
numpy
python-dotenv